reorganize to beautify all files
