+++ src/lib/dom.js
@@ -5,8 +5,8 @@
  * @param tag Tag to create
  * @param namespace Optional namespace value
  */
-exports.newEl = function (tag, namespace) {
-	if (!global.document) return;
+exports.newEl = function(tag, namespace) {
+    if (!global.document) return;
 
     if (namespace == null) {
         return document.createElement(tag);
@@ -22,7 +22,7 @@ exports.newEl = function (tag, namespace) {
  * @param el Reference to DOM element
  * @param attrs Object with attribute keys and values
  */
-exports.setAttr = function (el, attrs) {
+exports.setAttr = function(el, attrs) {
     for (var a in attrs) {
         el.setAttribute(a, attrs[a]);
     }
@@ -32,7 +32,7 @@ exports.setAttr = function (el, attrs) {
  * Creates a XML document
  * @private
  */
-exports.createXML = function () {
+exports.createXML = function() {
     if (!global.DOMParser) return;
     return new DOMParser().parseFromString('<xml />', 'application/xml');
 };
